extracting plugin manager to be a separate class
